Lift and shift core extension and npm packages (#5)
Copy extensions and packages directory from AzDO repo.

I want to maintain the same folder structure for now because I think it will be easy to "merge" from AzDO until we are ready to remove those folders from AzDO.

There was 1 small modification I had to make to the AccountProvider.unit.test.ts. Not sure why this was necessary (or rather why it worked in AzDO), but the compiler was complaining about Buffer.from overloads. The change was this:

<        const bufferFromStub = sinon.stub(Buffer, 'from').callsFake((data, encoding) => {
>        const bufferFromStub = sinon.stub(Buffer, 'from').callsFake((...args: [any, any?]) => {
>            const [data, encoding] = args;

You should be able to see it in the individual commits below, but its not very clean because I had commented out those lines to get build working