Loose ends from 558a670 (#22)
Update the commit ids used in migration to reflect which commits to use as a baseline.
This was always planned after the migration.

What was unplanned was that publishing the npm packages to the internal feed did not work:

```
npm error code EPRIVATE
npm error This package has been marked as private
npm error Remove the 'private' field from the package.json to publish it.
```

The package was marked as "private", so updated package.json to get rid of the private attribute