Add better error message when there is an API version mismatch (#45)
When there is an API version mismatch between the core and satellite
extension, the core extension will refuse to add the satellite. This
results in an error message that is unclear how to address the issue:

> Text (unlocalized): "Activating extension
'fabric.vscode-fabric-functions' failed: Extension
fabric.vscode-fabric-functions is not compatible with API version 0.8."

- - -

This change adds more information to the message: 

> Text (localized): "Activating extension 'fabric.vscode-fabric-functions'
failed: Extension fabric.vscode-fabric-functions (version 0.31.41-pre)
is not compatible with Microsoft Fabric extension (version 0.35.42). Use
the latest release versions for both extensions."